Use tabular font for numeric fields
When you’re entering numbers into a field we already add some extra
tracking to make it easier to read.

With this extra tracking the kerning looks a bit more even with the
tabular, not lining numbers. This makes sense because tabular numbers
are designed to be used where the content is numeric-only. Lining
numbers (the default) are more appropriate for numbers that are used in
passages of text.