Bump diff-dom to latest without Object.entries
Versions past this make use of Object.entries
which is not supported in IE11 (so causes an
error). This is meant to be taken care of by
polyfills but our frontend build pipeline isn't
set up to polyfill NPM modules at the moment.

This moves the version bump to the latest that
will still work with our current pipeline.