Remove peer dependencies
Peer dependencies are defined by projects intended
to be used by other projects (ie, libraries).

This doesn't apply to this project so removing the
peer dependencies field.

I'm also not sure why the `request` dependency was
there as it isn't a peer dependency of any of the
packages that use it in this project's dependency
tree.