Fix rendering of scrollable tables
The scrollable tables code styles some of the cells in the target table
by looking for the `table-field-center-aligned` class.

This class was renamed in 0512f40ad3d7fdb4d285a4f7e44659cca6f3e09f

This commit updates the scrollable tables code to refer to the new
classname, which means that things should line up properly when drawing
the table.