Handle table being wrapped in a `<details>` tag
If the table is wrapped in a `<details>` tag when
the page loads, it is removed from the layout.
Because of this, calculations of position and
dimension go wrong.

This adds some code to deal with this by opening
the tag while calculations happen, adding it to
the layout for that period.