Force dimensions to reset when <details> opens
The table is wrapped in a frame* to allow it to
scroll horizontally without scrolling the page.

The height of this frame is made to match the
amount of it that is visible in the window.
Because this can change while its parent
`<details>` tag is closed means the height may not
be correctly reported as it is not in the flow of
the document.

This adds a call to `maintainHeight` when it opens
to force it to recalculate the height when it is
re-inserted into the layout.

*Actually a `<div>` with `overflow-x: auto`
applied.