Add button from GOVUK Frontend to dashboard
Intended to test that a GOVUK Frontend can
successfully render in our app.

Note: the code is slightly different from that in
the example on the GOV.UK Design System:

https://design-system.service.gov.uk/components/button/

The example is a JavaScript function, called with
an object literal whereas our code is a Python
function called with a dictionary so the key needs
to be quoted.